Moved download_list_adapter from being a member of DownloadListView.Widgets class to DownloadListView itself.
Widgets was supposed to be passed on to DownloadListView constructor by
MainActivity.

Currently, it is designed so that DownloadListView initializes
DownloadListViewAdapter using its own Map<Long,DownloadListViewData> entries, and
its supplied instance of IViewRequestsHandler.

TODO: Figure out is it some way better or worse to pass DownloadListViewAdapter
also as a constructor parameter.