Entry:   Added @NonNull annotations for id and item. AbstractDownloadListView, AbstractDownloader:   Changed Entry<DataClass> add_xxxxyyyy(DataClass data) to           void             add_xxxxyyyy(Entry<DataClass> entry)   Rationale:     Delegating `id` generation to the Main class will help main class     keep track of related pieces of information that refers to a single download entry.     DownloadListView has a Map<Long,DownloadListViewData> while     Downloader will probably have a Map<Long, DownloadInfo>.     So if the Main class could inject the same id for both of them, it doesn't have     to keep a Map<Long,Long> mainId_DownloadInfoId and Map<Long,Long> mainId_DownloadListViewId     to track different pieces of same download entry.
It has been designed this way so that the different parts do not mess up
data to be used by other parts without notifying them.
According to current design, there are some specific channels of information flow,
which ensures that all data are in sync between different sections of the app,
even though they are highly decoupled.